Improve SSE hooks
